Current WIP tika requires data.content for sqs search
The sqs search is not abstracted through the API.  New internals for
Apache Tika search result in slightly differently stored data.  As
such we now send the sqs to data.content instead.

This is sufficient for current tests but may be removed in the future
in case a backwards compatible API is offered.